Removed many warnings
git-svn-id: http://isdb1.offis.uni-oldenburg.de/repos/odysseus/trunk@11013 aa0bda67-78ed-0310-a382-c5338b0f89ba

Former-commit-id: 13f709b4a4519440faacad16bcbbe49066cc0336

+++ benchmark/benchmarker/src/de/uniol/inf/is/odysseus/benchmarker/DescriptiveStatistics.java
@@ -89,7 +89,7 @@ public class DescriptiveStatistics {
 		double dev = 0.0;
 		@SuppressWarnings("unused")
 		double accum2 = 0.0;
-		this.mean = ((double) this.sum) / this.count;
+		this.mean = this.sum / this.count;
 		Arrays.sort(this.values);
 		this.offset = this.values.length - this.count;
 		if (offset == this.values.length) {
